Move DTO into directories
