+++ src/Response/DTO/Collection/ListImages.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
-namespace AstrobinWs\Response\DTO;
+namespace AstrobinWs\Response\DTO\Collection;
 
+use AstrobinWs\Response\DTO\AstrobinResponse;
+use AstrobinWs\Response\DTO\Item\Image;
 use AstrobinWs\Response\Iterators\ImageIterator;
 
 /**
