+++ src/Response/DTO/Item/Today.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace AstrobinWs\Response\DTO;
+namespace AstrobinWs\Response\DTO\Item;
 
 use AstrobinWs\Response\AbstractResponse;
-use AstrobinWs\Response\Iterators\ImageIterator;
+use AstrobinWs\Response\DTO\AstrobinResponse;
 
 /**
  * Class Today
