+++ src/Services/GetImage.php
@@ -8,8 +8,8 @@ use AstrobinWs\Exceptions\WsResponseException;
 use AstrobinWs\Filters\ImageFilters;
 use AstrobinWs\Filters\QueryFilters;
 use AstrobinWs\Response\DTO\AstrobinResponse;
-use AstrobinWs\Response\DTO\Image;
-use AstrobinWs\Response\DTO\ListImages;
+use AstrobinWs\Response\DTO\Collection\ListImages;
+use AstrobinWs\Response\DTO\Item\Image;
 use DateTime;
 use JsonException;
 use ReflectionException;
