+++ tests/Services/GetTodayImageTest.php
@@ -5,9 +5,9 @@ namespace Services;
 use AstrobinWs\Exceptions\WsException;
 use AstrobinWs\Exceptions\WsResponseException;
 use AstrobinWs\Response\DTO\AstrobinError;
-use AstrobinWs\Response\DTO\Image;
-use AstrobinWs\Response\DTO\ListToday;
-use AstrobinWs\Response\DTO\Today;
+use AstrobinWs\Response\DTO\Collection\ListToday;
+use AstrobinWs\Response\DTO\Item\Image;
+use AstrobinWs\Response\DTO\Item\Today;
 use AstrobinWs\Services\GetTodayImage;
 use Exception;
 use JsonException;
